Update to latest libnyquist. Squash some size_t int conversion warnings

+++ include/LabSound/core/AudioParam.h
@@ -66,7 +66,7 @@ public:
     
     // Calculates numberOfValues parameter values starting at the context's current time.
     // Must be called in the context's render thread.
-    void calculateSampleAccurateValues(ContextRenderLock&, float* values, unsigned numberOfValues);
+    void calculateSampleAccurateValues(ContextRenderLock&, float* values, size_t numberOfValues);
 
     // Connect an audio-rate signal to control this parameter.
     static void connect(ContextGraphLock& g, std::shared_ptr<AudioParam>, std::shared_ptr<AudioNodeOutput>);
